requirements.txt
typer[all]==0.15.1
GitPython==3.1.43
python-dateutil==2.9.0
holidays==0.59
